Allow nesting of Quips during Site creation
There's a circular dependency when a nested child object tries to verify its
parent during creation. The parent has been validated, but not yet saved, so it
doesn't have an ID yet. The child will complain that its parent can't be blank.

Work around this by adding a :nested attribute to Quips that gets set during a
nested creation. In this case, it doesn't try to validate the Site.