Track User creation for GA funnel
`Users#create` is straightforward - make a new `User`, log the event.

On `Users#authorize`, there are three possible states:
  * `User` is logged in and adds an `Authorization`: no new account
  * `User` is not logged in but `Authorization` exists: no new account
  * `User` is not logged in and `Authorization` is new: new account

The not-logged-in cases depend on `viewer` being `nil`, but stubbing it out
breaks the assignmend upon login, so the `before` action is just a safety to
ensure that nothing has accidentally set `viewer`.